fix: implement isolate request IDs, timeout recovery, and memory safety
- Add request ID tracking to SmartTurnIsolate to correctly correlate worker responses
- Add isolate timeout handling and automatic worker restart on timeout
- Support inferenceTimeoutMs and optional logger callback in SmartTurnConfig
- Ensure all native allocations in SmartTurnOnnxSession are properly cleaned up in try/finally blocks
- Fix stub/web SmartTurnOnnxSession return types and dynamic library resolution
- Fix audio preprocessor byte alignment validation and update doc comments

+++ pipecat_smart_turn_platform_interface/lib/src/onnx_inference_web.dart
@@ -7,8 +7,10 @@ import 'package:pipecat_smart_turn_platform_interface/src/exceptions.dart';
 import 'package:pipecat_smart_turn_platform_interface/src/mel_spectrogram.dart';
 import 'package:pipecat_smart_turn_platform_interface/src/onnx_runtime_web.dart';
 
+/// Resolves ONNX library path (web implementation returns null).
 String? resolveOnnxLibraryPath() => null;
 
+/// Resolves bundled model path for web asset loading.
 Future<String> extractBundledModel() async {
   return 'assets/$kDefaultModelAssetPath';
 }
@@ -76,7 +78,7 @@ class SmartTurnOnnxSession {
 
       // Model outputs a single probability tensor 'logits' of shape [batch, 1].
       final logitsTensor = outputs['logits'];
-      final probability = logitsTensor.data.toDart[0] as double;
+      final probability = logitsTensor.data.toDart[0];
 
       return probability;
     } on Object catch (e) {
